+++ CHANGELOG.md
@@ -16,3 +16,4 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `restart_rx` method
 - `set_rssi_offset` method
 - `set_rssi_collision_threshold`, `set_rssi_threshold` methods
+- `set_afc_bw` method